Lab 2 with other arch

+++ backend/src/main/java/com/example/confectionery/service/UserService.java
@@ -1,8 +1,12 @@
 package com.example.confectionery.service;
 
+import com.example.confectionery.dto.UserRegisterRequest;
+import com.example.confectionery.dto.UserResponse;
+import com.example.confectionery.entity.Role;
 import com.example.confectionery.entity.User;
 import com.example.confectionery.exception.ResourceNotFoundException;
 import com.example.confectionery.exception.UserAlreadyExistsException;
+import com.example.confectionery.mapper.UserResponseMapper;
 import com.example.confectionery.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,30 +19,45 @@ import java.util.List;
 public class UserService {
 
     private final UserRepository userRepository;
+    private final UserResponseMapper userResponseMapper;
 
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
+    public List<UserResponse> getAllUsers() {
+        return userRepository.findAll()
+                .stream()
+                .map(userResponseMapper)
+                .toList();
     }
 
-    public User getUserById(Long id) {
+    public UserResponse getUserById(Long id) {
         return userRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Пользователь с ID " + id + " не найден"));
+                .map(userResponseMapper)
+                .orElseThrow(() -> new ResourceNotFoundException("Пользователь не найден"));
     }
 
     @Transactional
-    public User createUser(User user) {
-        if (userRepository.existsByEmail(user.getEmail())) {
+    public UserResponse createUser(UserRegisterRequest request) {
+        if (userRepository.existsByEmail(request.getEmail())) {
             throw new UserAlreadyExistsException("Пользователь с таким email уже существует");
         }
 
-        if (user.getRole() == null) {
-            user.setRole(User.Role.USER);
-        }
-        return userRepository.save(user);
+        User userEntity = User.builder()
+                .firstName(request.getFirstName())
+                .lastName(request.getLastName())
+                .email(request.getEmail())
+                .password(request.getPassword())
+                .phone(request.getPhone())
+                .role(Role.USER)
+                .build();
+
+        User savedUser = userRepository.save(userEntity);
+        return userResponseMapper.apply(savedUser);
     }
 
     @Transactional
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Пользователь не найден");
+        }
         userRepository.deleteById(id);
     }
 }
